moving functions from note_et_avis to analyzer
