+++ src/food_analysis/core/analyzer.py
@@ -3,61 +3,74 @@
 Version simple pour démarrer. L'équipe pourra ajouter plus de méthodes.
 """
 
-import matplotlib.pyplot as plt
 import pandas as pd
 
+# Fonction pour calculer les statistiques des recettes
 
-class DataAnalyzer:
-    """Classe pour analyser les données de recettes et interactions."""
 
-    # def __init__(self, recipes: pd.DataFrame, interactions: pd.DataFrame) -> None:
-    #     """Initialise l'analyseur avec les données."""
-    #     self.recipes = recipes
-    #     self.interactions = interactions
+def compute_recipe_stats(
+    recipe_df: pd.DataFrame, interaction_df: pd.DataFrame, m: int = 10
+) -> pd.DataFrame:
+    """
+    Calcule la note moyenne, le nombre d'avis et la note pondérée pour chaque recette.
 
-    # def get_basic_stats(self) -> Dict[str, Any]:
-    #     """Retourne des statistiques de base sur les données."""
-    #     stats = {
-    #         "total_recipes": len(self.recipes),
-    #         "total_interactions": len(self.interactions),
-    #         "total_users": self.interactions["user_id"].nunique()
-    #         if "user_id" in self.interactions.columns
-    #         else 0,
-    #     }
-    #     return stats
+    Args:
+        recipe_df (pd.DataFrame): DataFrame des recettes
+        interaction_df (pd.DataFrame): DataFrame des interactions
+        m (int): Nombre minimal d'avis pour la pondération
 
-    # def get_top_recipes(self, n: int = 10) -> pd.DataFrame:
-    #     """Retourne les N recettes avec le plus d'interactions."""
-    #     if self.interactions.empty:
-    #         return pd.DataFrame()
+    Returns:
+        pd.DataFrame: DataFrame avec id, nom, avg_rating, n_reviews et weighted_rating
+    """
+    # Calculer la note moyenne et le nombre d'avis par recette
+    recipe_stats = (
+        interaction_df.groupby("recipe_id")
+        .agg(avg_rating=("rating", "mean"), n_reviews=("rating", "count"))
+        .reset_index()
+    )
 
-    #     recipe_counts = self.interactions["recipe_id"].value_counts().head(n)
-    #     top_recipes = self.recipes[self.recipes["id"].isin(recipe_counts.index)]
-    #     return top_recipes
+    # Note moyenne globale
+    C = recipe_stats["avg_rating"].mean()
 
-    # def get_average_rating(self) -> float:
-    #     """Calcule la note moyenne de toutes les recettes."""
-    #     if "rating" not in self.interactions.columns:
-    #         return 0.0
+    # Calcul de la note pondérée
+    recipe_stats["weighted_rating"] = (
+        recipe_stats["n_reviews"] / (recipe_stats["n_reviews"] + m)
+    ) * recipe_stats["avg_rating"] + (m / (recipe_stats["n_reviews"] + m)) * C
 
-    #     valid_ratings = self.interactions[self.interactions["rating"] > 0]["rating"]
-    #     if len(valid_ratings) == 0:
-    #         return 0.0
+    # Fusion avec le DataFrame recipe pour récupérer le nom
+    recipe_stats_with_name = pd.merge(
+        recipe_stats,
+        recipe_df[["id", "name"]],
+        left_on="recipe_id",
+        right_on="id",
+        how="left",
+    )
 
-    #     return float(valid_ratings.mean())
+    # Trier par note pondérée décroissante
+    recipe_stats_with_name = recipe_stats_with_name.sort_values(
+        "weighted_rating", ascending=False
+    ).reset_index(drop=True)
 
-    def plot_rating_distribution(
-        self, interaction_df: pd.DataFrame, recipe_id: int
-    ) -> None:
-        """
-        Affiche la distribution des notes pour une recette spécifique.
-        """
+    return recipe_stats_with_name[
+        ["name", "avg_rating", "n_reviews", "weighted_rating"]
+    ]
 
-        ratings = interaction_df.loc[interaction_df["recipe_id"] == recipe_id, "rating"]
-        plt.hist(ratings, bins=6, edgecolor="black")
-        plt.xlim(-0.25, None)
-        plt.title(f"Distribution des notes pour la recette {recipe_id}")
-        plt.xlabel("Note")
-        plt.ylabel("Nombre d'avis")
-        # plt.xticks([])
-        plt.show()
+
+def recipe_reviews(recipe_id: int, interaction_df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Récupère les avis pour une recette donnée.
+
+    Args:
+        recipe_id (int): ID de la recette
+        interaction_df (pd.DataFrame): DataFrame des interactions
+
+    Returns:
+        pd.DataFrame: DataFrame contenant les avis pour la recette
+    """
+    return (
+        interaction_df[interaction_df["recipe_id"] == recipe_id][
+            ["user_id", "rating", "date", "review"]
+        ]
+        .sort_values("date", ascending=False)
+        .reset_index(drop=True)
+    )
